Preliminary support for EPUB3 output
The real change is that now there are 2 different ADOC_TOTALs, one for PDF
and one for EPUB. This is because the workarounds currently in place for
frontmatter styling don't work in EPUB output (because an EPUB can't
"include" a PDF).

At the moment, both of the ADOC_TOTALs are identical and the frontmatter bug
is still present in EPUB output, but at least the plumbing is in place to
make them different.