New approach for dealing with unmatched double quotes.
The old solution had a subtle, unwanted side-effect which didn't account for
closed double quotes butted against a nonword character (e.g. em-dash or
ellipsis). This approach is much better. It delegates the handling of curly
double quotes to Asciidoctor as much as possible *except for* the singular
times where we need to make an exception.